readyup|refactor: Place kit-discovery tests beside the code they cover
Tests for the repository sweep, the directory walker, and the recursive listing sit beside the sources they exercise, matching the layout the rest of the repository uses.

Helpers in the listing formatter are collected into one block at the end of the file, and comments that defended a design choice or restated the code below them are gone.
