+++ packages/readyup/.gitignore
@@ -3,6 +3,5 @@
 .readyup/kits/*.js
 .readyup/manifest.json
 
-# Anchored to the package root: unanchored, this would also swallow `src/schemas/` and
-# `__tests__/schemas/`, where a newly added file would go untracked without a word.
+# Anchored to the package root, so that `src/schemas/` and `__tests__/schemas/` stay tracked.
 /schemas/
