+++ packages/readyup/src/portable/__tests__/walkDirectories.unit.test.ts
@@ -8,8 +8,8 @@ vi.mock(import('node:fs'), async (importOriginal) => {
   return { ...actual, readdirSync: mockReaddirSync };
 });
 
-import { walkDirectories } from '../../src/portable/walkDirectories.ts';
-import { useTempDir } from '../helpers/tempDir.ts';
+import { useTempDir } from '../../../__tests__/helpers/tempDir.ts';
+import { walkDirectories } from '../walkDirectories.ts';
 
 const { readdirSync: readdirSyncActual } = await vi.importActual<typeof import('node:fs')>('node:fs');
 
@@ -27,21 +27,6 @@ const tempDir = useTempDir({
     tempDir.write('deep/one/two/package.json', '{}');
   },
 });
-
-/** The read options the sweep passes, which a passthrough hands straight back to the real reader. */
-interface SweepReaddirOptions {
-  encoding: 'utf8';
-  withFileTypes: true;
-}
-
-/** Fails the directory at `relativePath`, letting every other read through to the filesystem. */
-function failReadOf(relativePath: string, code: string): void {
-  const failingDir = `${tempDir.dir}/${relativePath}`;
-  mockReaddirSync.mockImplementation((dir: string, options: SweepReaddirOptions) => {
-    if (dir === failingDir) throw Object.assign(new Error(`read failed: ${code}`), { code });
-    return readdirSyncActual(dir, options);
-  });
-}
 
 describe(walkDirectories, () => {
   beforeEach(() => {
@@ -90,7 +75,6 @@ describe(walkDirectories, () => {
     expect(found).not.toContain('.hidden');
   });
 
-  // Pruning is the only thing that excludes anything, so a custom list reaches what the defaults hid.
   it('honors custom prune globs in place of the defaults', () => {
     const found = walkDirectories({ root: tempDir.dir, match: '**/package.json', prune: ['**/packages'] });
 
@@ -123,10 +107,28 @@ describe(walkDirectories, () => {
     expect(found).toStrictEqual([]);
   });
 
-  // An incomplete sweep must not pass for a complete one, so a systemic failure is not swallowed.
   it('rethrows a filesystem failure that is not benign', () => {
     failReadOf('packages/b', 'EMFILE');
 
     expect(() => walkDirectories({ root: tempDir.dir, match: '**/package.json' })).toThrow('read failed: EMFILE');
   });
 });
+
+// region | Helpers
+
+/** The read options the sweep passes, which a passthrough hands straight back to the real reader. */
+interface SweepReaddirOptions {
+  encoding: 'utf8';
+  withFileTypes: true;
+}
+
+/** Fails the directory at `relativePath`, letting every other read through to the filesystem. */
+function failReadOf(relativePath: string, code: string): void {
+  const failingDir = `${tempDir.dir}/${relativePath}`;
+  mockReaddirSync.mockImplementation((dir: string, options: SweepReaddirOptions) => {
+    if (dir === failingDir) throw Object.assign(new Error(`read failed: ${code}`), { code });
+    return readdirSyncActual(dir, options);
+  });
+}
+
+// endregion | Helpers
